Prevent halting before reboot if RTT logging is enabled
If the RTT logging backend is used and a JLink probe is connected, the
DHCSR register is set which indicates to the memfault sdk that a
debugger is attached, even though only one-way logging is used.
Execution halts at a breakpoint and the reboot is never reached.

Check whether the RTT backend is enabled and skip the breakpoint in this
case.

This is a temporary fix which might be solved by a dedicated config
option upstream.